fix: исправление замечаний по ревью

+++ src/delivery_service/api/schemas.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from decimal import Decimal
+
 from pydantic import BaseModel, Field
 
 
@@ -7,7 +9,7 @@ class PackageCreate(BaseModel):
     name: str = Field(min_length=1, max_length=200)
     weight: float = Field(gt=0)  # кг
     package_type_id: int = Field(gt=0)
-    content_value_usd: float = Field(ge=0)
+    content_value_usd: Decimal = Field(ge=0)
 
 
 class PackageCreatedResponse(BaseModel):
@@ -25,8 +27,8 @@ class PackageOut(BaseModel):
     weight: float
     package_type_id: int
     package_type_name: str
-    content_value_usd: float
-    delivery_cost_rub: float | None
+    content_value_usd: Decimal
+    delivery_cost_rub: Decimal | None
     delivery_cost: str
